Redesign meta decorator implementation
Previous implementation failed taking into account that meta had to be stored on a "class" level, meaning that the decorated "target" was incorrectly used in the internal registry. Now, with help from the builtin initialization logic from the decorator context, the target class can be obtained. Furthermore, this also solves the "field" decorator issue (I hope)...