Deprecate reflections components
The idea might be good, but the implementation and use-case of this will not yield desired results. Every class or method will require the reflect() decorator, where the pseudo reflection metadata must be processed. This will surely cause a performance overhead... For now, these components must be deprecated and re-designed to a more likable solution - if at all possible.