Remove experimental reflect decorator
The @reflect() decorator, along with all related components are now no longer needed. The idea was "good", but very cumbersome and possible too performance demanding. The new @targetMeta() decorator has taken the best out of this experiment and can now be used to associate metadata directly with class methods, which is good enough for now. Developers can use it to store their own "reflection-like" data as metadata, if needed.