Change logic of how application determines if it has booted or not
Previously, the service registrar's amount of booted service providers was used to determine if the application has been booted or not. However, in an edge-case, an application might not have any service providers registered / booted, before the application itself is attempted booted. This could lead to a false negative. Therefore, the application now has its own booted state.